support listening for both servers and clients.
most of the plumbing for these things is shared, they just want to
do different setup for the fd once it's been accepted. this factors
out the differences into a function pointer than the parser sets.

i dont have a client to test with yet, so this just tweaks the
server listener that's already there.